Cast time comparison as ints

+++ chopper.cpp
@@ -55,7 +55,7 @@ static const uint64_t maxtime = (1UL << 43);
 static const uint64_t maxjump = 10*50000000; // 50 MHz time
 
 // Maximum time drift allowed between two clocks without a complaint
-static const uint64_t maxdrift = 5000; // 50 MHz ticks (1 us)
+static const int maxdrift = 5000; // 50 MHz ticks (1 us)
 
 // This function writes macro files needed to correctly interpret the
 // chopped files with RAT.  It can be suppressed with the -t flag.
@@ -349,7 +349,7 @@ static void compute_times(const PmtEventRecord * const hits,
                      + hits->TriggerCardData.Bc10_1;
 
     // Check for consistency between clocks
-    const uint64_t dd = std::abs((oldtime10 - time10)*5 - (oldtime50 - time50));
+    const int dd = std::abs(int((oldtime10 - time10)*5) - int(oldtime50 - time50));
     if (dd > maxdrift){
       fprintf(stderr, "ALARM: The Clocks jumped by %i ticks!\n", dd);
     }
